feat: add configurability and filetype action (#1)
